requirements.txt
logzero
python-telegram-bot
tinydb